add stats for rank recipients question
